requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
firebase-admin
google-cloud-storage
python-dotenv
kafka-python